Made a better file structure
Now we will be using the MVVM architecture, to organize our files and classes
