+++ OctoGym/SwiftUIView2.swift
@@ -1,0 +1,18 @@
+//
+//  SwiftUIView2.swift
+//  OctoGym
+//
+//  Created by Olivia on 6/14/25.
+//
+
+import SwiftUI
+
+struct SwiftUIView2: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    SwiftUIView2()
+}
