+++ OctoGym/Views/ContentView.swift
@@ -1,0 +1,29 @@
+//
+//  ContentView.swift
+//  OctoGym
+//
+//  Created by Olivia on 6/14/25.
+//
+
+
+import SwiftUI
+
+struct ContentView: View {
+    @StateObject private var authViewModel = AuthViewModel()
+    
+    var body: some View {
+        Group {
+            if authViewModel.isLoggedIn {
+                HomeView()
+                    .environmentObject(authViewModel)
+            } else {
+                LoginView()
+                    .environmentObject(authViewModel)
+            }
+        }
+    }
+}
+
+#Preview {
+    ContentView()
+}
